Add rootdir redirect to main directory
The orginal template was designed for the .Rmd files to reside in the project directory. The organization scheme was updated to put .Rmd files in an 'rmarkdown' directory. However, the template should 'act like' it is in the main directory. Adding the 'knitr::rootdir...' to the template makes this the default behavior. A note was added to comment or delete this line, if the template is used in the project directory.