Fix Issue #3 Header Markdown bug
After upgrading knitr and RStudio, I noticed that the header syntax now needs to be separated from the title text (e.g. ###Conclusion### needs to be ### Conclusion ###). Otherwise, the ### is output explicitly instead of the markdown generating the appropriate html header. Therefore, need to ensure that all header text is separated from the # header syntax in the template.